Fix CI: use NXS_FLAG_V13_COMPACT_MASK in C and Short bitwise in Kotlin.
Include stdio.h for nxs_format_v13_compact_err snprintf in nxs.h.

+++ kotlin/src/main/kotlin/nxs/NxsReader.kt
@@ -68,7 +68,7 @@ class NxsReader
             dictHash = buf.getLong(8)
             tailPtr = buf.getLong(16)
             if (flags.toInt() and FLAG_V13_COMPACT_MASK != 0) {
-                val bits = flags and 0x01F0
+                val bits = flags.toInt() and 0x01F0
                 val hex = bits.toString(16).padStart(4, '0')
                 val msg =
                     "this file uses NXS v1.3 compact encoding (flags 0x$hex); " +
